Rename and copy all comments from sync.Map
